Adding blacklist for files with _ and whitelisting ending in whitelist. Adding gevent and concurrent requests. Skipping getting content from files over 1mb. Changing from waitress to Gunicorn

+++ webhook/views.py
@@ -1,8 +1,10 @@
+import engine
+import gevent
+import os
+
 from pyramid.response import Response
 from pyramid.view import view_config
 from pyramid.view import notfound_view_config
-import os
-import engine
 from restapi import RESTAPIService
 
 DEFAULTS = {
@@ -21,11 +23,30 @@ class RootView(object):
         else:
             self.api = api_service
 
-    def send_data(self, git_url, push_url, git_info):
+    def _push_to_api(self, push_url, url_list, headers):
+        """"""
+
+        # Unpack data
+        git_url = DEFAULTS['git_url']
+        git_branch = DEFAULTS['git_branch']
+
+        # Get content from github concurrently
+        requests =  self.api.get(git_url, urls=url_list, params=git_branch)
+
+        # Decode contents concurrently with Gevent
+        jobs = [gevent.spawn(engine.get_github_json, r) for r in requests]
+        gevent.joinall(jobs, timeout=2)
+
+        json_list = [job.value for job in jobs]
+        
+        # Put to API concurrently 
+        self.api.put(push_url, url_list, json_list, headers=headers)
+
+    def send_data(self, push_url, git_info):
         """This method formats and sends data to our API."""
+
         # Unpacking data
         data        = git_info['changes']
-        git_branch  = git_info['push_branch']
         base_url    = git_info['base_url']
         serving_url = git_info['serving_url']
 
@@ -36,27 +57,26 @@ class RootView(object):
         headers = {'content-type'   : u'application/json',
                    'repository_url' : base_url,
                    'authorization'  : bearer_token,
+                   'serving_url'    : serving_url,
                   }
-        # add data
-        for add_data in data[0]:
-            headers['serving_url'] = serving_url + add_data
-            raw_data_from_github = self.api.get(git_url+add_data, params=git_branch)
-            github_json = engine.get_github_json(raw_data_from_github)
-            self.api.put(push_url+add_data, github_json, headers=headers)
-        # update data
-        for update_data in data[1]:
+        # Add Data
+        if data[0]:
+            add_urls = engine.create_async_lists_by_structure(data[0])
+            self._push_to_api(push_url, add_urls, headers)
+        # Update data
+        if data[1]:
+            update_urls = engine.create_async_lists_by_structure(data[1])
             headers['info'] = u'updated'
-            headers['serving_url'] = serving_url + update_data
-            raw_data_from_github = self.api.get(git_url+update_data, params=git_branch)
-            github_json = engine.get_github_json(raw_data_from_github)
-            self.api.put(push_url+update_data, github_json, headers=headers)
-        # remove data
-        for delete_data in data[2]:
-            self.api.delete(push_url+delete_data)
+            self._push_to_api(push_url, update_urls, headers)
+        # Delete data
+        if data[2]:
+            for delete_data in data[2]:
+                self.api.delete(push_url+delete_data)
 
     @view_config(route_name='root', request_method='POST')
     def default_view(self):
         """Main view, receives webhooks from github and sends to configured API."""
+
         # check for ENV variables
         if DEFAULTS['git_url'] == None:
             return Response(u'GITHUB_WEBHOOK_GIT_URL not set', status=400)
@@ -66,7 +86,6 @@ class RootView(object):
             return Response(u'GITHUB_WEBHOOK_PUSH_URL not set', status=400)
 
         # We have all the env variables, unpack data
-        git_url         = DEFAULTS['git_url']
         git_branch      = DEFAULTS['git_branch']
         push_url        = DEFAULTS['push_url']
         use_github_auth = DEFAULTS['use_github_auth']
@@ -94,21 +113,20 @@ class RootView(object):
         git_info['base_url']    = engine.get_base_url(data)
         git_info['serving_url'] = engine.get_serving_url(git_info['base_url'])
         git_info['author']      = engine.get_author(data)
-        git_info['push_branch'] = engine.get_branch(data)
         git_info['changes']     = engine.get_changes(data)
 
         # Checking if the branch is right and checking if the API answers
-        if git_branch == git_info['push_branch']:
+        if git_branch == engine.get_branch(data):
             r = self.api.get(push_url)
             if r.status_code == 200:
                 # Everything is OK, sending data to API
-                self.send_data(git_url, push_url, git_info)
+                self.send_data(push_url, git_info)
                 return Response(u'{0}\nSuccessfuly commited to {1}'.format(git_info['author'],git_branch))
             else:
-                return Response(u'Failed to connect to API\nStatus:' + str(r.status_code), status=403)
+                return Response(u'Failed to connect to API\nStatus:' + str(r.status_code), status=504)
         else:
-            response_msg = u'{0} wrong branch!\nYou commited to {1}.\nOnly accepting commits to {2} branch.'.format(git_info['author'], git_info['push_branch'], git_branch)
-            return Response(response_msg, status=504)
+            response_msg = u'{0} wrong branch!\nOnly accepting commits to {1} branch.'.format(git_info['author'], git_branch)
+            return Response(response_msg, status=403)
 
 @notfound_view_config(request_method='GET')
 def not_found_view(self):
